+++ packages-internal/utils/src/index.ts
@@ -2,4 +2,5 @@ export { TokenService } from './services/token-service/token-service';
 export { getExchangeRates } from './services/pricing-service/exchange-rates';
 export { addressItem, textItem, currencyItem, nodeIDItem, dataItem, dateItem } from './utils/detail-item';
 export { retry } from './utils/retry';
+export { fetchAndVerify } from './utils/fetch-and-verify';
 export { getCoreHeaders } from './utils/get-core-headers';